api/requirements.txt
Flask
supabase
openai
python-dotenv
